Added orig_nbits to header (#35)
Scott added a new header entry orig_nbits to parfaits_utils and this is used to determine the original number of bits used in the data written. I believe in our case, not filling this variable made psrfits_utils think that the data has 0 bits and converts everything written to zeroes.