Improves performance of beamforming kernel.
The task of forming a beam is assigned to a single warp instead of
an entire block of threads for better utilisation and the ability
of threads within a warp to efficiently cooperate.

The reduction (sum) is now performed in parallel across the warp instead
of being performed sequentially by thread 0.