Reduces context switches by creating just enough blocks.
In this implementation each beam computation, one for each frequency channel
and time sample, is assigned to a warp (32/64 consecutive threads working in
lockstep). A warp is assigned one or more beams to compute depending on how
many thread blocks (hence warps) are created. This number is now unrelated
to the problem at hand and depends on the hardware specifics for better
performance. We create "just enough" blocks to avoid too many context
switches.